basic clone/fetch/push functionality works
with this commit, we are now able to push a repository into ipfs
and clone a repository from ipfs.  This will only be mildly functional
until it is possible to publish to arbitrary names using ipns, because
without that feature it's not possible to "update" the repository in
ipfs in way that will be visible (the exception being that you could
publish a *single* repository using your node name and possibly use an
ipns reference for that, although I haven't tested that feature yet).